Move the game loop into CGameLoop
